+++ core/io/io.h
@@ -26,7 +26,7 @@ SOFTWARE.
 
 extern void ASM_OUTB(unsigned int port, unsigned int data);
 extern void ASM_OUTL(unsigned int port, unsigned int data);
-extern unsigned short ASM_INB(unsigned short port);
+extern unsigned int ASM_INB(unsigned int port);
 extern unsigned int ASM_INL(unsigned short port);
 
 extern void ASM_INSW(unsigned int port, unsigned int numWords, unsigned int buffer);
